# ngo-rest-api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "OpenAPI Specification for NGO",
        "description": "OpenAPI Specification for NGO"
    },
    "host": "localhost:10010",
    "basePath": "/api",
    "paths": {
...
        "/Donation": {
            "post": {
                "tags": [
                    "Donation"
                ],
                "summary": "Create a new instance of the model and persist it into the data source.",
                "operationId": "Donation.create",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Model instance data",
                        "required": false,
                        "schema": {
                            "description": "Model instance data",
                            "$ref": "#/definitions/Donation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "$ref": "#/definitions/Donation"
                        }
                    }
                },
                "deprecated": false
            },
            "get": {
                "tags": [
                    "Donation"
                ],
                "summary": "Find all instances of the model matched by filter from the data source.",
                "operationId": "Donation.find",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter defining fields, where, include, order, offset, and limit - must be a JSON-encoded string ({\"something\":\"value\"})",
                        "required": false,
                        "type": "string",
                        "format": "JSON"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Donation"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/Donation/{id}": {
            "head": {
                "tags": [
                    "Donation"
                ],
                "summary": "Check whether a model instance exists in the data source.",
                "operationId": "Donation.exists",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Model id",
                        "required": true,
                        "type": "string",
                        "format": "JSON"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "exists": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "deprecated": false
            },
            "get": {
                "tags": [
                    "Donation"
                ],
                "summary": "Find a model instance by {{id}} from the data source.",
                "operationId": "Donation.findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Model id",
                        "required": true,
                        "type": "string",
                        "format": "JSON"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter defining fields and include - must be a JSON-encoded string ({\"something\":\"value\"})",
                        "required": false,
                        "type": "string",
                        "format": "JSON"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "$ref": "#/definitions/Donation"
                        }
                    }
                },
                "deprecated": false
            },
            "put": {
                "tags": [
                    "Donation"
                ],
                "summary": "Replace attributes for a model instance and persist it into the data source.",
                "operationId": "Donation.replaceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Model id",
                        "required": true,
                        "type": "string",
                        "format": "JSON"
                    },
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Model instance data",
                        "required": false,
                        "schema": {
                            "description": "Model instance data",
                            "$ref": "#/definitions/Donation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "$ref": "#/definitions/Donation"
                        }
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": [
                    "Donation"
                ],
                "summary": "Delete a model instance by {{id}} from the data source.",
                "operationId": "Donation.deleteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Model id",
                        "required": true,
                        "type": "string",
                        "format": "JSON"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "type": "object"
                        }
                    }
                },
                "deprecated": false
            }
        },
...
        "/NGO": {
            "post": {
                "tags": [
                    "NGO"
                ],
                "summary": "Create a new instance of the model and persist it into the data source.",
                "operationId": "NGO.create",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Model instance data",
                        "required": false,
                        "schema": {
                            "description": "Model instance data",
                            "$ref": "#/definitions/NGO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "$ref": "#/definitions/NGO"
                        }
                    }
                },
                "deprecated": false
            },
            "get": {
                "tags": [
                    "NGO"
                ],
                "summary": "Find all instances of the model matched by filter from the data source.",
                "operationId": "NGO.find",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter defining fields, where, include, order, offset, and limit - must be a JSON-encoded string ({\"something\":\"value\"})",
                        "required": false,
                        "type": "string",
                        "format": "JSON"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/NGO"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
        "/Donor": {
            "post": {
                "tags": [
                    "Donor"
                ],
                "summary": "Create a new instance of the model and persist it into the data source.",
                "operationId": "Donor.create",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Model instance data",
                        "required": false,
                        "schema": {
                            "description": "Model instance data",
                            "$ref": "#/definitions/Donor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "$ref": "#/definitions/Donor"
                        }
                    }
                },
                "deprecated": false
            },
            "get": {
                "tags": [
                    "Donor"
                ],
                "summary": "Find all instances of the model matched by filter from the data source.",
                "operationId": "Donor.find",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter defining fields, where, include, order, offset, and limit - must be a JSON-encoded string ({\"something\":\"value\"})",
                        "required": false,
                        "type": "string",
                        "format": "JSON"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request was successful",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Donor"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
...
    "consumes": [
        "application/json",
        "application/x-www-form-urlencoded",
        "application/xml",
        "text/xml"
    ],
    "produces": [
        "application/json",
        "application/xml",
        "text/xml",
        "application/javascript",
        "text/javascript"
    ],
    "definitions": {
        "Donation": {
            "description": "An asset named Donation",
            "properties": {
                "donationId": {
                    "description": "The instance identifier for this type",
                    "type": "string"
                },
                "donationAmount": {
                    "type": "number",
                    "format": "double"
                },
                "donationDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "donor": {
                    "description": "The identifier of an instance of donor",
                    "$ref": "#/definitions/x-any"
                },
                "ngo": {
                    "description": "The identifier of an instance of ngo",
                    "$ref": "#/definitions/x-any"
                }
            },
            "required": [
                "donationId",
                "donationAmount",
                "donationDate",
                "donor",
                "ngo"
            ],
            "additionalProperties": false
        },
...
        "Donor": {
            "description": "A participant named Donor",
            "properties": {
                "$class": {
                    "default": "Donor",
                    "description": "The class identifier for this type",
                    "type": "string"
                },
                "donorUserName": {
                    "description": "The instance identifier for this type",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "registeredDate": {
                    "type": "string",
                    "format": "date-time"
                }
            },
            "required": [
                "donorUserName",
                "email",
                "registeredDate"
            ],
            "additionalProperties": false
        },
...
    }
}